former : refactoring parsing
